CSS change in the nav-bar (#189)
* CSS change in the nav-bar 

Added a design  to the <a> tags of .nav-links class. This is just to make the UI look more awesome :D .

* Update brackets.io.css

* CSS change in the nav-bar

Added the required formatting to the CSS code and moved animation to within the nav-bar

* Update brackets.io.css

* Removed the blank line

I was unable to understand earlier which blank line to remove. I guess this is the one you wanted.

* Removed the empty line

* The last nits